+++ Src/App/Commands/Interfaces/IAccountInteractionCommand.cs
@@ -1,0 +1,9 @@
+﻿using BMBank.Src.DatabaseInteraction.Core.DAO;
+
+namespace BMBank.Src.App.Commands.Interfaces
+{
+    public interface IAccountInteractionCommand
+    {
+        public AccountDAO DAO { get; }
+    }
+}