+++ Src/DatabaseInteraction/Connection/IConnectionFactory.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Data.SqlClient;
+
+namespace BookOrg.Src.Logic.Connection
+{
+    /// <include file='../../../Docs/ClassDocumentation.xml' path='ClassDocumentation/ClassMembers[@name="IConnectionFactory"]/*'/>
+    public interface IConnectionFactory
+    {
+        SqlConnection? CreateConnection();
+    }
+}